refactor query execution in preparation for implementing optimizations
specifically, now Exec returns an Executor object on which multiple
slices can be provided in a streaming fashion, and Executor can be reused
until it is closed by caller. besides initializing executor at a given
node, most of the query execution is also now consolidated in the Run
function of query Executor.