fix memory usage for video outputs
This reverts label buffers to be single-reader again, where they consume
the labels so that the labels are not stuck stored in the buffer. To
provide multi-reader (e.g. for Exec where we may want to save the label,
output the label, and also wait for execution to complete), buffer now
offers a Split function that can split its output across multiple other
buffers.